move git url parsing to github pkg
