+++ github/url_test.go
@@ -1,0 +1,154 @@
+package github
+
+import (
+	"net/url"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"gotest.tools/v3/assert"
+)
+
+func TestParseGitPseudoURLSuccess(t *testing.T) {
+	testCases := []struct {
+		name   string
+		inURL  string
+		wantGU GitURL
+	}{
+		{
+			name:  "valid SSH URL",
+			inURL: "git@github.com:Pix4D/cogito.git",
+			wantGU: GitURL{
+				URL: &url.URL{
+					Scheme: "ssh",
+					User:   url.User("git"),
+					Host:   "github.com",
+					Path:   "/Pix4D/cogito.git",
+				},
+				Owner: "Pix4D",
+				Repo:  "cogito",
+			},
+		},
+		{
+			name:  "valid HTTPS URL",
+			inURL: "https://github.com/Pix4D/cogito.git",
+			wantGU: GitURL{
+				URL: &url.URL{
+					Scheme: "https",
+					Host:   "github.com",
+					Path:   "/Pix4D/cogito.git",
+				},
+				Owner: "Pix4D",
+				Repo:  "cogito",
+			},
+		},
+		{
+			name:  "valid HTTP URL",
+			inURL: "http://github.com/Pix4D/cogito.git",
+			wantGU: GitURL{
+				URL: &url.URL{
+					Scheme: "http",
+					Host:   "github.com",
+					Path:   "/Pix4D/cogito.git",
+				},
+				Owner: "Pix4D",
+				Repo:  "cogito",
+			},
+		},
+		{
+			name:  "valid HTTPS URL with username:password",
+			inURL: "https://username:password@github.com/Pix4D/cogito.git",
+			wantGU: GitURL{
+				URL: &url.URL{
+					Scheme: "https",
+					User:   url.UserPassword("username", "password"),
+					Host:   "github.com",
+					Path:   "/Pix4D/cogito.git",
+				},
+				Owner: "Pix4D",
+				Repo:  "cogito",
+			},
+		},
+		{
+			name:  "valid HTTP URL with username:password",
+			inURL: "http://username:password@github.com/Pix4D/cogito.git",
+			wantGU: GitURL{
+				URL: &url.URL{
+					Scheme: "http",
+					User:   url.UserPassword("username", "password"),
+					Host:   "github.com",
+					Path:   "/Pix4D/cogito.git",
+				},
+				Owner: "Pix4D",
+				Repo:  "cogito",
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			gitUrl, err := ParseGitPseudoURL(tc.inURL)
+
+			if err != nil {
+				t.Fatalf("\nhave: %s\nwant: <no error>", err)
+			}
+			if diff := cmp.Diff(tc.wantGU, gitUrl, cmp.Comparer(
+				func(x, y *url.Userinfo) bool {
+					return x.String() == y.String()
+				})); diff != "" {
+				t.Errorf("gitURL: (-want +have):\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestParseGitPseudoURLFailure(t *testing.T) {
+	testCases := []struct {
+		name    string
+		inURL   string
+		wantErr string
+	}{
+		{
+			name:    "totally invalid URL",
+			inURL:   "hello",
+			wantErr: "invalid git URL hello: missing scheme",
+		},
+		{
+			name:    "invalid SSH URL",
+			inURL:   "git@github.com/Pix4D/cogito.git",
+			wantErr: "invalid git SSH URL git@github.com/Pix4D/cogito.git: want exactly one ':'",
+		},
+		{
+			name:    "invalid HTTPS URL",
+			inURL:   "https://github.com:Pix4D/cogito.git",
+			wantErr: `invalid port ":Pix4D" after host`,
+		},
+		{
+			name:    "invalid HTTP URL",
+			inURL:   "http://github.com:Pix4D/cogito.git",
+			wantErr: `invalid port ":Pix4D" after host`,
+		},
+		{
+			name:    "too few path components",
+			inURL:   "http://github.com/cogito.git",
+			wantErr: "invalid git URL: path: want: 3 components; have: 2 [ cogito.git]",
+		},
+		{
+			name:    "too many path components",
+			inURL:   "http://github.com/1/2/cogito.git",
+			wantErr: "invalid git URL: path: want: 3 components; have: 4 [ 1 2 cogito.git]",
+		},
+		{
+			name:    "No leaked password in invalid URL with username:password",
+			inURL:   "http://username:password@github.com/Pix4D/cogito.git\n",
+			wantErr: `net/url: invalid control character in URL`,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			_, err := ParseGitPseudoURL(tc.inURL)
+
+			assert.Error(t, err, tc.wantErr)
+		})
+	}
+}
